refactor(tui): extract SessionTab model from statusview
SessionTab is a full model that owns the rendering of a single session
tab (name, workspace, windows, accent bar, background). The parent
statusview manages a map of tabs keyed by session ID and syncs them
via Update when sessions, windows, selection, or focus change.
